requirements.txt
alembic==1.11.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cffi==1.16.0
click==8.1.6
cryptography==41.0.5
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-engineio==4.5.1
python-socketio==5.8.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.1.1
